chore: refactor the codebase | #8
