--- api/config/db.js
@@ -1,15 +1,0 @@
-import mongoose from "mongoose";
-
-const connectDB = async () => {
-  try {
-    await mongoose.connect(process.env.MONGO_URI, {
-     
-    });
-    console.log("✅ MongoDB Connected");
-  } catch (error) {
-    console.error("❌ MongoDB Connection Failed:", error);
-    process.exit(1);
-  }
-};
-
-export default connectDB;
