+++ api/src/app.js
@@ -1,17 +1,14 @@
-// app.js - Main application setup
 import express from 'express';
 import cors from 'cors';
 import path from 'path';
 import { fileURLToPath } from 'url';
 import connectDB from './config/database.js';
 import apiRoutes from './routes/api.js';
-import indexRoute from './routes/index.js';
 import errorHandler from './middleware/errorHandler.js';
 import spaceRoutes from './routes/spaceRoutes.js';
 import chatRoutes from './routes/chatRoutes.js';
-import queryRoutes from './routes/queryRoutes.js'; // Add this import
-// ...
- 
+import queryRoutes from './routes/queryRoutes.js';
+
 // Connect to MongoDB
 connectDB();
 
@@ -25,7 +22,6 @@ app.use('/api/spaces', spaceRoutes);
 app.use('/api/chat', chatRoutes);
 app.use('/api', queryRoutes); // Add this line for query routes
 app.use('/api', apiRoutes);
-app.use('/', indexRoute);
 
 // Serve React app in production
 if (process.env.NODE_ENV === 'production') {
