+++ api/src/config/pinecone.js
