+++ api/src/services/embeddingService.js
