+++ api/src/services/pineconeService.js
