--- api/views/index.html
@@ -1,161 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>PDF RAG App with Pinecone</title>
-  <style>
-    body { font-family: Arial, sans-serif; margin: 20px; }
-    .container { max-width: 800px; margin: 0 auto; }
-    .card { border: 1px solid #ddd; padding: 20px; margin-bottom: 20px; border-radius: 5px; }
-    input, button, textarea { margin: 10px 0; padding: 8px; width: 100%; }
-    button { background-color: #4CAF50; color: white; border: none; cursor: pointer; }
-    #documents { margin-top: 20px; }
-    .document-item { padding: 10px; border-bottom: 1px solid #eee; cursor: pointer; }
-    .document-item:hover { background-color: #f5f5f5; }
-  </style>
-</head>
-<body>
-  <div class="container">
-    <h1>PDF RAG Application with Pinecone</h1>
-    
-    <div class="card">
-      <h2>Upload a PDF</h2>
-      <form id="uploadForm">
-        <input type="file" id="pdfFile" accept=".pdf" required>
-        <button type="submit">Upload & Process</button>
-      </form>
-      <div id="uploadStatus"></div>
-    </div>
-    
-    <div class="card">
-      <h2>Your Documents</h2>
-      <div id="documents">Loading...</div>
-    </div>
-    
-    <div class="card">
-      <h2>Ask Questions</h2>
-      <div>
-        <p>Selected Document: <span id="selectedDoc">None</span></p>
-        <textarea id="queryInput" placeholder="Ask a question about the selected document..." rows="3"></textarea>
-        <button id="queryBtn" disabled>Ask</button>
-        <div id="queryResult"></div>
-      </div>
-    </div>
-  </div>
-
-  <script>
-    let selectedDocId = null;
-
-    // Load documents on page load
-    fetchDocuments();
-
-    document.getElementById('uploadForm').addEventListener('submit', async (e) => {
-      e.preventDefault();
-      const fileInput = document.getElementById('pdfFile');
-      const status = document.getElementById('uploadStatus');
-      
-      if (!fileInput.files[0]) {
-        status.textContent = 'Please select a file.';
-        return;
-      }
-      
-      const formData = new FormData();
-      formData.append('pdf', fileInput.files[0]);
-      
-      status.textContent = 'Uploading and processing...';
-      
-      try {
-        const response = await fetch('/api/upload', {
-          method: 'POST',
-          body: formData
-        });
-        
-        const result = await response.json();
-        if (response.ok) {
-          status.textContent = 'Document processed successfully!';
-          fileInput.value = '';
-          fetchDocuments();
-        } else {
-          status.textContent = 'Error: ' + result.error;
-        }
-      } catch (error) {
-        status.textContent = 'Upload failed: ' + error.message;
-      }
-    });
-
-    document.getElementById('queryBtn').addEventListener('click', async () => {
-      const query = document.getElementById('queryInput').value.trim();
-      const resultDiv = document.getElementById('queryResult');
-      
-      if (!query) {
-        resultDiv.textContent = 'Please enter a question.';
-        return;
-      }
-      
-      if (!selectedDocId) {
-        resultDiv.textContent = 'Please select a document first.';
-        return;
-      }
-      
-      resultDiv.textContent = 'Processing query...';
-      
-      try {
-        const response = await fetch('/api/query', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json'
-          },
-          body: JSON.stringify({
-            documentId: selectedDocId,
-            query: query
-          })
-        });
-        
-        const result = await response.json();
-        if (response.ok) {
-          resultDiv.innerHTML = '<strong>Answer:</strong><br>' + result.answer;
-        } else {
-          resultDiv.textContent = 'Error: ' + result.error;
-        }
-      } catch (error) {
-        resultDiv.textContent = 'Query failed: ' + error.message;
-      }
-    });
-
-    async function fetchDocuments() {
-      const docsDiv = document.getElementById('documents');
-      
-      try {
-        const response = await fetch('/api/documents');
-        const documents = await response.json();
-        
-        if (documents.length === 0) {
-          docsDiv.textContent = 'No documents yet. Upload one to get started.';
-          return;
-        }
-        
-        docsDiv.innerHTML = '';
-        documents.forEach(doc => {
-          const docEl = document.createElement('div');
-          docEl.className = 'document-item';
-          docEl.textContent = doc.title;
-          docEl.addEventListener('click', () => {
-            // Highlight selected document
-            document.querySelectorAll('.document-item').forEach(el => {
-              el.style.backgroundColor = '';
-            });
-            docEl.style.backgroundColor = '#e0f7e0';
-            
-            // Set as selected document
-            selectedDocId = doc._id;
-            document.getElementById('selectedDoc').textContent = doc.title;
-            document.getElementById('queryBtn').disabled = false;
-          });
-          docsDiv.appendChild(docEl);
-        });
-      } catch (error) {
-        docsDiv.textContent = 'Failed to load documents: ' + error.message;
-      }
-    }
-  </script>
-</body>
-</html>
